requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.8.2
branca==0.7.2
cattrs==23.2.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.1
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
folium==0.17.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.2
MarkupSafe==2.1.5
multidict==6.0.5
numpy==1.26.4
openmeteo_requests==1.2.0
openmeteo_sdk==1.11.7
pandas==2.2.2
platformdirs==4.2.2
py4j==0.10.9.7
pydantic==1.10.16
pyspark==3.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
redis==5.0.6
requests==2.32.3
requests-cache==1.2.0
retry-requests==2.0.0
simple-websocket==1.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
url-normalize==1.4.3
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
xyzservices==2024.6.0
yarl==1.9.4
